+++ Base/BaseMod/Base/BaseWorldGen.cs
@@ -38,7 +38,7 @@ namespace AAMod
 			if (heightLimit == -1) heightLimit = (int)Main.worldSurface;
 			if(amountInWorld == -1)
 			{
-				float oreCount = (float)(Main.maxTilesX / 4200);
+				float oreCount = Main.maxTilesX / 4200;
 				oreCount *= 50f;
 				amountInWorld = (int)oreCount;
 			}
@@ -84,7 +84,7 @@ namespace AAMod
             for(int y = startY; y < Main.maxTilesY - 10; y++)
             {
                 Tile tile = Framing.GetTileSafely(x, y);
-                if(tile != null && tile.nactive() && (!solid || Main.tileSolid[(int)tile.type])){ return y; }
+                if(tile != null && tile.nactive() && (!solid || Main.tileSolid[tile.type])){ return y; }
             }
             return Main.maxTilesY - 10;
         }
@@ -100,7 +100,7 @@ namespace AAMod
             for (int y = startY; y > 10; y--)
             {
                 Tile tile = Framing.GetTileSafely(x, y);
-                if (tile != null && tile.nactive() && (!solid || Main.tileSolid[(int)tile.type])) { return y; }
+                if (tile != null && tile.nactive() && (!solid || Main.tileSolid[tile.type])) { return y; }
             }
             return 10;
         }
@@ -114,7 +114,7 @@ namespace AAMod
 				for(int x = startX; x > 10; x--)
 				{
 					Tile tile = Framing.GetTileSafely(x, y);
-					if(tile != null && tile.nactive() && (!solid || Main.tileSolid[(int)tile.type])){ return x; }
+					if(tile != null && tile.nactive() && (!solid || Main.tileSolid[tile.type])){ return x; }
 				}			
 				return 10;
 			}else
@@ -122,7 +122,7 @@ namespace AAMod
 				for(int x = startX; x < Main.maxTilesX - 10; x++)
 				{
 					Tile tile = Framing.GetTileSafely(x, y);
-					if(tile != null && tile.nactive() && (!solid || Main.tileSolid[(int)tile.type])){ return x; }
+					if(tile != null && tile.nactive() && (!solid || Main.tileSolid[tile.type])){ return x; }
 				}
 				return Main.maxTilesX - 10;				
 			}
@@ -161,10 +161,10 @@ namespace AAMod
          */
         public static void ReplaceTiles(Vector2 position, int radius, int[] tiles, int[] replacements, bool silent = false, bool sync = true)
         {
-            int radiusLeft = (int)(position.X / 16f - (float)radius);
-            int radiusRight = (int)(position.X / 16f + (float)radius);
-            int radiusUp = (int)(position.Y / 16f - (float)radius);
-            int radiusDown = (int)(position.Y / 16f + (float)radius);
+            int radiusLeft = (int)(position.X / 16f - radius);
+            int radiusRight = (int)(position.X / 16f + radius);
+            int radiusUp = (int)(position.Y / 16f - radius);
+            int radiusDown = (int)(position.Y / 16f + radius);
             if(radiusLeft < 0){ radiusLeft = 0; } if(radiusRight > Main.maxTilesX) { radiusRight = Main.maxTilesX; }
             if(radiusUp < 0){ radiusUp = 0; } if(radiusDown > Main.maxTilesY) { radiusDown = Main.maxTilesY; }
 
@@ -362,8 +362,8 @@ namespace AAMod
             }
             if (sync && Main.netMode != 0)
             {
-                int sizeWidth = tileWidth + (int)Math.Max(0, (width - 1));
-                int sizeHeight = tileHeight + (int)Math.Max(0, (height - 1));
+                int sizeWidth = tileWidth + Math.Max(0, (width - 1));
+                int sizeHeight = tileHeight + Math.Max(0, (height - 1));
                 int size = sizeWidth > sizeHeight ? sizeWidth : sizeHeight;
                 NetMessage.SendTileSquare(-1, x + (int)(size * 0.5F), y + (int)(size * 0.5F), size + 1);
             }
@@ -425,7 +425,7 @@ namespace AAMod
 				{
 					int size = ((endY - y) > (endX - x) ? (endY - y) : (endX - x));
 					if (thickness > size) size = thickness;
-					NetMessage.SendData(20, -1, -1, NetworkText.FromLiteral(""), size, (float)x, (float)y, 0f, 0);
+					NetMessage.SendData(20, -1, -1, NetworkText.FromLiteral(""), size, x, y, 0f, 0);
 				}
 			}else //genning a line that isn't straight
 			{
@@ -434,7 +434,7 @@ namespace AAMod
 				float length = Vector2.Distance(start, end);
 				float way = 0f;
 
-				float rot = (float)BaseUtility.RotationTo(start, end); if(rot < 0f) rot = (float)(Math.PI * 2f) - Math.Abs(rot);
+				float rot = BaseUtility.RotationTo(start, end); if(rot < 0f) rot = (float)(Math.PI * 2f) - Math.Abs(rot);
 				float rotPercent = MathHelper.Lerp(0f, 1f, rot / (float)(Math.PI * 2f));
 				bool horizontal = rotPercent < 0.125f || (rotPercent > 0.375f && rotPercent < 0.625f) || rotPercent > 0.825f;
 				int tileIndex = -1, wallIndex = -1;
@@ -463,7 +463,7 @@ namespace AAMod
 					}
 					if (sync && Main.netMode != 0 && ((!horizontal && Math.Abs(lastY - point.Y) >= 5) || (horizontal && Math.Abs(lastY - point.Y) >= 5) || (way + 1 > length)))
 					{
-						int size = (int)Math.Max(5, thickness);
+						int size = Math.Max(5, thickness);
 						NetMessage.SendData(10, -1, -1, NetworkText.FromLiteral(""), lastX, lastY, size, size, 0);
 						lastX = point.X; lastY = point.Y;
 					}
@@ -725,7 +725,7 @@ namespace AAMod
 							{
 								if (Main.tile[x + 1, y].type != 190 && Main.tile[x + 1, y].type != 48 && Main.tile[x + 1, y].type != 232 && WorldGen.SolidTile(x - 1, y + 1) && WorldGen.SolidTile(x + 1, y) && !Main.tile[x - 1, y].active() && !Main.tile[x + 1, y - 1].active())
 								{
-									WorldGen.PlaceTile(x, y, (int)Main.tile[x, y + 1].type, false, false, -1, 0);
+									WorldGen.PlaceTile(x, y, Main.tile[x, y + 1].type, false, false, -1, 0);
 									if (WorldGen.genRand.Next(2) == 0)
 									{
 										WorldGen.SlopeTile(x, y, 2);
@@ -737,7 +737,7 @@ namespace AAMod
 								}
 								if (Main.tile[x - 1, y].type != 190 && Main.tile[x - 1, y].type != 48 && Main.tile[x - 1, y].type != 232 && WorldGen.SolidTile(x + 1, y + 1) && WorldGen.SolidTile(x - 1, y) && !Main.tile[x + 1, y].active() && !Main.tile[x - 1, y - 1].active())
 								{
-									WorldGen.PlaceTile(x, y, (int)Main.tile[x, y + 1].type, false, false, -1, 0);
+									WorldGen.PlaceTile(x, y, Main.tile[x, y + 1].type, false, false, -1, 0);
 									if (WorldGen.genRand.Next(2) == 0)
 									{
 										WorldGen.SlopeTile(x, y, 1);
@@ -946,10 +946,10 @@ namespace AAMod
 			Point trueOrigin = origin;
 			for(int m = 0; m < length; m++)
 			{
-				int height = (int)MathHelper.Lerp((float)startheight, (float)endheight, (float)m / (float)length);
+				int height = (int)MathHelper.Lerp(startheight, endheight, m / (float)length);
 				if(heightVariance != null)
 				{
-					height = Math.Max(endheight, (int)((float)startheight * BaseUtility.MultiLerp((float)m / (float)length, heightVariance)));
+					height = Math.Max(endheight, (int)(startheight * BaseUtility.MultiLerp(m / (float)length, heightVariance)));
 				}
 				int x = trueOrigin.X + (dir ? m : -m);
 				int y = trueOrigin.Y + (startheight - height);
@@ -1014,10 +1014,10 @@ namespace AAMod
 			Point trueOrigin = origin;
 			for(int m = 0; m < depth; m++)
 			{
-				int width = (int)MathHelper.Lerp((float)startwidth, (float)endwidth, (float)m / (float)depth);
+				int width = (int)MathHelper.Lerp(startwidth, endwidth, m / (float)depth);
 				if(widthVariance != null)
 				{
-					width = Math.Max(endwidth, (int)((float)startwidth * BaseUtility.MultiLerp((float)m / (float)depth, widthVariance)));
+					width = Math.Max(endwidth, (int)(startwidth * BaseUtility.MultiLerp(m / (float)depth, widthVariance)));
 				}
 				int x = trueOrigin.X + (startwidth - width);
 				int y = trueOrigin.Y + (dir ? m : -m);
@@ -1123,7 +1123,7 @@ namespace AAMod
 		{
 			if(x < 0 || x > Main.maxTilesX || y < 0 || y > Main.maxTilesY) return false;
 			if(GenBase._tiles[x, y] == null) GenBase._tiles[x, y] = new Tile();
-			Main.Map.UpdateLighting(x, y, (byte)Math.Max(Main.Map[x, y].Light, _brightness));
+			Main.Map.UpdateLighting(x, y, Math.Max(Main.Map[x, y].Light, _brightness));
 			return base.UnitApply(origin, x, y, args);
 		}
 	}	
@@ -1182,11 +1182,11 @@ namespace AAMod
 
 		public override bool Apply(Point origin, int x, int y, params object[] args)
 		{
-			Vector2 value = new Vector2((float)origin.X + (_width / 2), (float)origin.Y);
-			Vector2 value2 = new Vector2((float)x, (float)y);
+			Vector2 value = new Vector2((float)origin.X + (_width / 2), origin.Y);
+			Vector2 value2 = new Vector2(x, y);
 			float num = Vector2.Distance(value2, value);
 			float num2 = Math.Max(0f, Math.Min(1f, (num - this._innerRadius) / (this._outerRadius - this._innerRadius)));
-			if (GenBase._random.NextDouble() > (double)num2)
+			if (GenBase._random.NextDouble() > num2)
 			{
 				return base.UnitApply(origin, x, y, args);
 			}
